third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/container-type-snap-parsing.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Containment Test: Parsing of container-type with snap</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-type">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/css-contain/container-queries/support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
  setup(() => assert_implements_container_queries());

  test_valid_value('container-type', 'snap');
  test_valid_value('container-type', 'size snap');
  test_valid_value('container-type', 'snap inline-size', 'inline-size snap');
  test_valid_value('container-type', 'snap sticky', 'sticky snap');
  test_valid_value('container-type', 'size sticky snap');

  test_invalid_value('container-type', 'snap snap');
  test_invalid_value('container-type', 'normal snap');
  test_invalid_value('container-type', 'inline-size block-size snap');
</script>
